Hake inject and make_fuzz_target work with multi-file projects

+++ test_rust/make_fuzz_targets.py
@@ -1,18 +1,14 @@
 #!/usr/bin/env python3
 
 from pathlib import Path
+from posixpath import relpath
 import shutil
 import os
 import re
+import sys
 
 import tomlkit
 
-fuzz_common = Path("fuzz_common").absolute()
-bug_collection = Path("bug_collection").absolute()
-bug_fuzzing_collection = Path("bug_fuzzing_collection").absolute()
-
-
-bug_fuzzing_collection.mkdir()
 
 main_def_pattern = re.compile(r'fn\s*main')
 main_pub_def_pattern = re.compile(r'pub\s*fn\s*main')
@@ -22,41 +18,67 @@ def make_main_public(src):
     else:
         return main_def_pattern.sub("pub fn main", src)
 
-cargo_src_f = Path("app/Cargo.toml")
-cargo_dest = ""
-with cargo_src_f.open("r") as f:
-    cargo_src = tomlkit.load(f)
+def main(argv):
+    src_app = Path("app/").resolve()
+    if len(argv) >= 2:
+        src_app = Path(argv[1]).resolve()
 
-    # replace bin target by lib target
-    # TODO: keep both bin and lib target to make testing easier
-    cargo_src.pop("bin")
+    fuzz_common = Path("fuzz_common").resolve()
+    bug_collection = Path("bug_collection").resolve()
+    bug_fuzzing_collection = Path("bug_fuzzing_collection").resolve()
 
-    if not "lib" in cargo_src:
-        cargo_src["lib"] = {}
+    bug_fuzzing_collection.mkdir()
 
-    cargo_dest = tomlkit.dumps(cargo_src)
+    cargo_src_f = src_app / "Cargo.toml"
+    cargo_dest_content = ""
+    with cargo_src_f.open("r") as f:
+        cargo_src = tomlkit.load(f)
 
+        if not "lib" in cargo_src:
+            cargo_src["lib"] = {}
 
-for i, bug_src in enumerate(bug_collection.iterdir()):
-    print(f"Processing bug {bug_src}")
-    destdir = bug_fuzzing_collection / f"bug{i}"
-    destdir.mkdir()
-
-    # write bug injection parameters to a file
-    (destdir / "injection_description.txt").write_text(bug_src.stem)
-
-    # copy source code to target
-    # additionnaly replace "fn main" by "pub fn main"
-
-    codedir = destdir / "src"
-    codedir.mkdir()
-    dest_code = codedir / "lib.rs"
-    dest_code.write_text(make_main_public(bug_src.read_text()))
-
-    # write a cargo file
-    (destdir / "Cargo.toml").write_text(cargo_dest)
+        cargo_dest_content = tomlkit.dumps(cargo_src)
 
 
-    # symlink the fuzz targets
-    fuzz_dir = destdir / "fuzz"
-    os.symlink(os.path.relpath(fuzz_common, destdir), fuzz_dir)
+    for bug_src in bug_collection.iterdir():
+        print(f"Processing bug {bug_src}")
+
+        bugfile = bug_src / "bug.rs"
+        infofile = bug_src / "info.txt"
+
+        # read the original path from the info file
+        info_lines = infofile.read_text().splitlines()
+        orig_path = Path(info_lines[0]).resolve()
+        bugfile_relpath = orig_path.relative_to(src_app)
+
+        destdir = bug_fuzzing_collection / bug_src.name
+        if (destdir.exists()):
+            print("WARN: skipping this bug, target dir {destir} exists already.")
+            continue
+        
+        # replicate the app, bug change the content to
+        # 1. contain the file with the injected bug
+        # 2. make main public (if an executable)
+        # 3. expose a lib target in Cargo.toml (if not present)
+        shutil.copytree(src_app, destdir)
+
+        # keep track of info
+        shutil.copy(infofile, destdir)
+
+        # 1. Place file with injected bug
+        bugfile_in_dest = destdir / bugfile_relpath
+        shutil.copy(bugfile, bugfile_in_dest)
+        #bugfile_in_dest.write_text(make_main_public(bug_src.read_text()))
+
+        # write a cargo file
+        (destdir / "Cargo.toml").write_text(cargo_dest_content)
+
+
+        # symlink the fuzz targets
+        fuzz_dir = destdir / "fuzz"
+        os.symlink(os.path.relpath(fuzz_common, destdir), fuzz_dir)
+
+
+
+if __name__ == "__main__":
+    main(sys.argv)
